Move RPCPool from `package cluster` => `package pools`
